requirement.txt
# Django
Django==6.0

# Database (if using SQLite, it's built-in, else use psycopg2/boto3 if PostgreSQL/AWS)
# psycopg2-binary==2.9.9 

# ML / Deep Learning
tensorflow==2.14.0
keras==2.14.0
numpy==1.26.0
Pillow==10.0.0

# Blockchain
web3==6.0.0

# Misc
django-environ==0.10.0  # Optional, for environment variables
python-dotenv==1.0.0     # Optional, if using .env file

# For sending emails in development
django-anymail==9.3  

# Random utilities
requests==2.31.0
